requirements.txt
anyio==4.8.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Flask==3.1.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
python-telegram-bot==21.10
requests==2.32.3
sniffio==1.3.1
urllib3==2.3.0
Werkzeug==3.1.3
